# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LastATS - Applicant Tracking System API
servers:
  - url: /api
    description: Base API path
...
security:
  - clerkAuth: []
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications with filtering, sorting, and rating aggregation
      parameters:
        - $ref: "#/components/parameters/OrganizationIdHeader"
        - name: jobId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
        - name: search
          in: query
          description: Search by candidate name or email
          schema:
            type: string
        - name: minRating
          in: query
          description: Minimum average rating filter
          schema:
            type: string
        - name: dateFrom
          in: query
          description: Filter applications from this date (ISO 8601)
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: Filter applications up to this date (ISO 8601)
          schema:
            type: string
            format: date
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [appliedAt, candidateName, rating, status]
            default: appliedAt
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedApplications"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new application
      parameters:
        - $ref: "#/components/parameters/OrganizationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplication"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application details with ratings
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/OrganizationIdHeader"
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/OrganizationIdHeader"
      responses:
        "204":
          description: Application deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/OrganizationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/ApplicationStatus"
                notifyCandidate:
                  type: boolean
                  description: Whether to send an email notification to the candidate about the status change
              required:
                - status
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /applications/{id}/resume:
    get:
      operationId: getApplicationResume
      tags: [applications]
      summary: Stream the candidate resume file (auth via Bearer token only)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Resume file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /applications/{id}/notes:
    patch:
      operationId: updateApplicationNotes
      tags: [applications]
      summary: Update application internal notes
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/OrganizationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
              required:
                - notes
      responses:
        "200":
          description: Notes updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /applications/{id}/email:
    post:
      operationId: sendApplicationEmail
      tags: [applications]
      summary: Send an email to the candidate for this application
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/OrganizationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                templateId:
                  type: string
                  format: uuid
                subject:
                  type: string
                htmlBody:
                  type: string
                textBody:
                  type: string
              required:
                - subject
                - htmlBody
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  emailLogId:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /applications/{id}/emails:
    get:
      operationId: getApplicationEmails
      tags: [applications]
      summary: Get email history for an application
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/OrganizationIdHeader"
      responses:
        "200":
          description: Email history list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /applications/bulk-email:
    post:
      operationId: sendBulkEmail
      tags: [applications]
      summary: Send bulk email to selected candidates
      parameters:
        - $ref: "#/components/parameters/OrganizationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicationIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                templateId:
                  type: string
                  format: uuid
                subject:
                  type: string
                htmlBody:
                  type: string
                textBody:
                  type: string
              required:
                - applicationIds
                - subject
                - htmlBody
      responses:
        "200":
          description: Bulk email results
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  total:
                    type: integer
                  sent:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        applicationId:
                          type: string
                        emailLogId:
                          type: string
                        status:
                          type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token
  parameters:
...
    OrganizationIdHeader:
      name: X-Organization-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The organization context for this request
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    ApplicationStatus:
      type: string
      enum: [new, reviewed, shortlisted, rejected, hired]
...
    PaginatedApplications:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            allOf:
              - $ref: "#/components/schemas/Application"
              - type: object
                properties:
                  candidateFirstName:
                    type: string
                  candidateLastName:
                    type: string
                  candidateEmail:
                    type: string
                  candidatePhone:
                    type: string
                    nullable: true
                  candidateResumeUrl:
                    type: string
                    nullable: true
                  jobTitle:
                    type: string
                  jobDepartment:
                    type: string
                    nullable: true
                  avgRating:
                    type: string
                    nullable: true
                    description: Average rating as decimal string
                  ratingCount:
                    type: integer
                    nullable: true
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
...
    CreateApplication:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        candidateId:
          type: string
          format: uuid
        coverLetter:
          type: string
        customFieldResponses:
          type: object
          additionalProperties:
            type: string
      required:
        - jobId
        - candidateId
...